refactor(warp_runner): make separate function for main task
