+++ ui/src/main.rs
@@ -283,8 +283,8 @@ fn bootstrap(cx: Scope) -> Element {
     logger::trace("rendering bootstrap");
 
     // warp_runner must be started from within a tokio reactor
-    let mut warp_runner = warp_runner::WarpRunner::init();
-    warp_runner.run(WARP_EVENT_CH.tx.clone(), WARP_CMD_CH.rx.clone());
+    let mut warp_runner = warp_runner::WarpRunner::new();
+    warp_runner.run();
 
     // make the window smaller while the user authenticates
     let desktop = use_window(cx);
